requirements.txt
# Pin versions loosely; upgrade transformers if Gemma 4 load fails.
torch>=2.2.0
transformers>=4.49.0
accelerate>=0.34.0
safetensors>=0.4.0
Pillow>=10.0.0
streamlit>=1.38.0
